envoytun/requirements.txt
google-cloud-bigquery>=2.30.0
#google-cloud-secret-manager>=2.7.2
google-cloud-storage>=1.42.0